Add specific Sea class
