+++ src/environment.ts
@@ -1,37 +1,16 @@
 import * as THREE from "three";
-import { Water } from "three/addons/objects/Water.js";
 import { Sky } from "three/addons/objects/Sky.js";
 
+import { Sea } from "./sea";
+
 export class Environment {
-  private _water: Water;
   private _sky: Sky;
   private _pmremGenerator: THREE.PMREMGenerator;
   private _sceneEnv: THREE.Scene;
   private _renderTarget: THREE.WebGLRenderTarget | undefined;
+  private _sea: Sea;
 
   constructor(private scene: THREE.Scene, private renderer: THREE.WebGLRenderer) {
-    const waterGeometry = new THREE.PlaneGeometry(1000, 1000, 1000, 1000);
-
-    this._water = new Water(waterGeometry, {
-      textureWidth: 512,
-      textureHeight: 512,
-      waterNormals: new THREE.TextureLoader().load(
-        "textures/waternormals.jpg",
-        (texture: THREE.Texture) => {
-          texture.wrapS = texture.wrapT = THREE.RepeatWrapping;
-        },
-      ),
-      sunDirection: new THREE.Vector3(),
-      sunColor: 0xffffff,
-      waterColor: 0x001e0f,
-      distortionScale: 3.7,
-      fog: scene.fog !== undefined,
-    });
-
-    this._water.rotation.x = -Math.PI / 2;
-
-    scene.add(this._water);
-
     // Skybox
 
     this._sky = new Sky();
@@ -47,6 +26,8 @@ export class Environment {
 
     this._pmremGenerator = new THREE.PMREMGenerator(renderer);
     this._sceneEnv = new THREE.Scene();
+
+    this._sea = new Sea(scene);
   }
 
   animate(): void {
@@ -60,7 +41,6 @@ export class Environment {
     sun.setFromSphericalCoords(1, phi, theta);
 
     this._sky.material.uniforms.sunPosition.value.copy(sun);
-    this._water.material.uniforms.sunDirection.value.copy(sun).normalize();
 
     if (this._renderTarget !== undefined) {
       this._renderTarget.dispose();
@@ -72,20 +52,6 @@ export class Environment {
 
     this.scene.environment = this._renderTarget.texture;
 
-    const vertices = this._water.geometry.attributes.position.array;
-    const waveAngle = Math.PI / 9;
-    const cosAngle = Math.cos(waveAngle);
-    const sinAngle = Math.sin(waveAngle);
-
-    for (let i = 0; i < vertices.length; i += 3) {
-      const x = vertices[i];
-      const y = vertices[i + 1];
-      const period = 0.05;
-      const rotatedX = x * cosAngle - y * sinAngle;
-      vertices[i + 2] = Math.sin(rotatedX * period + time) * 4;
-    }
-
-    this._water.geometry.attributes.position.needsUpdate = true;
-    this._water.material.uniforms.time.value += 1.0 / 60.0;
+    this._sea.animate(time, sun);
   }
 }
